d s c donct

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: morgani <morgani@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/01/09 10:58:02 by vlecoq-v          #+#    #+#             */
-/*   Updated: 2019/01/17 13:55:50 by morgani          ###   ########.fr       */
+/*   Updated: 2019/01/17 14:29:41 by morgani          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -47,7 +47,7 @@ typedef struct		s_conv
 	int				sn;
 	char			*str;
 	long long		nbr;
-	long long		cmpt;
+	int				cmpt;
 	int				prt_spc;
 }					t_conv;
